server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-generativeai
PyMuPDF
python-docx
aiofiles